Fixed bug with icons in vaults page
